teleport nachbearbeitet jetzt kommt man wieder zurück

+++ windowFrame/src/game/maps/OverworldMap.java
@@ -54,7 +54,7 @@ public class OverworldMap extends Maps {
 	      hashMapMaps.put(new Point2D.Double(73,93) ,  overworldmap );
 	      hashMapMaps.put(new Point2D.Double(75,102) , overworldmap );
 	      hashMapMaps.put(new Point2D.Double(102,100) ,overworldmap );
-	      hashMapMaps.put(new Point2D.Double(98,90) ,  overworldmap );
+	      hashMapMaps.put(new Point2D.Double(98,92) ,  overworldmap );
 	      hashMapMaps.put(new Point2D.Double(104,64) , overworldmap );
 	      hashMapMaps.put(new Point2D.Double(165,34) , overworldmap );
 	      hashMapMaps.put(new Point2D.Double(93, 39) , overworldmap );
@@ -62,7 +62,7 @@ public class OverworldMap extends Maps {
 	      teleMap.put(new Point2D.Double(73,93)  , new Point2D.Double(75,103) );
 	      teleMap.put(new Point2D.Double(75,102) , new Point2D.Double(73,94)  );
 	      teleMap.put(new Point2D.Double(102,100), new Point2D.Double(98,93)  );
-	      teleMap.put(new Point2D.Double(98,90)  , new Point2D.Double(102,101));
+	      teleMap.put(new Point2D.Double(98,92)  , new Point2D.Double(102,101));
 	      teleMap.put(new Point2D.Double(104,64) , new Point2D.Double(93, 40) );
 	      teleMap.put(new Point2D.Double(165,34) , new Point2D.Double(104,65) );
 	      teleMap.put(new Point2D.Double(93, 39) , new Point2D.Double(157,36) );
